Perform a quick precheck before the regex search
It is a common good pattern in Moodle filters to firstly perform a quick
and cheap check for a keyword presence in the text before attempting to
do the full regex based search and replace. This significantly improves
the performance, especially where the filter syntax is quite rare in the
texts.

Such a check in this case effectively replaces the empty() and
is_numeric() prechecks.