Improve checks for parent languages
Following advice from David Mudrak, instead of manually handling the
parent language in the replace_callback() method, it would be better to
make use of string_manager::get_language_dependencies(). It returns the
whole chain of parent languages without the need to switch the current
language.

Thanks a lot to David for the suggestion.

Closes #4