Normalize languge pack names and do it faster
While the regexp for the filter allowed language names using either '-'
or '_' in their names, Moodle language packs always use '_'. So we need
to make sure that we change any '-' into '_'. Added a couple of unit
tests to check for this.

Also, we can use strtolower instead of core_text::strtolower() as
language short names are ASCII only, and strtolower is much, much
faster. We also don't need trim(), as the regex capture doesn't include
trailing/leading whitespace. This speeds up the filtering quite a bit.