Move parentcache initialization and updating to filter top function
Before we performed the operation in the callback for every lang block
that was matched by the filter. But the current language (and thus its
parent language(s)) didn't change between callback invocations. So we
were doing unneeded work.

Also make the parentcache a class static member, instead of making it
an instance member. That way we don't need to rebuilt it every time
the filter instance is created.