Fix breakage introduced in commit 8c29e07
Even if we had some unit tests in place, they were not sufficient to
catch the breakage introduced in commit 8c29e07. The behavior
introduced in that commit is not compatible with the intended behavior
of the fallback 'other' language tag. They behave identically only in
very simple cases (like some of the unit tests we had), but not in the
general case.

The root of the problem is that the callback function is called for
each langblock separately. So we don't know what happened in other
langblocks, and we can't decide whether to fallback to language
'other'. Thus, it is not possible to do a single
preg_replace_callback() operation. We need to do it once for the
user active language. And only after all the langblocks were processed we
can decide if we need to do it a second time (if no langblock matched)
for the 'other' language.

But we can at least avoid duplicating the replace callback function if
we refactor it to accept the filtering language, and call it with
either the user active language, or 'other'.

Closes #15