Refactor units tests to work in Moodle 3.11 and later.
Use filter_multilang as the basis for the refactoring.

Obviously this means that this plugin version should only be used in
Moodle 3.11 and later. Not because the plugin itself can't run in
earlier versions (it should be able to run on anything since Moodle
3.5). But because the unit tests will fail on anything earlier than
Moodle 3.11 (as the way unit tests are implemented changed in an
incompatible way in that release).

This is why we are bumping the plugin version to 2.0.0. To signal the
breaking change.

Also extract the data used for the actual unit test to a separate
file, as the idea is to keep two separate branches for the Moodle 3.10
and earlier versions, and Moodle 3.11 and later versions. But to reuse
the unit tests data across the two branches as easily as possible,
without having to touch other parts of the code.

Fixes: #32