Remove the "string" return type constraint from "filter" function
The filter function can actually return any data type. Because a) the
signature for the base class doesn't include a return type constraint,
and b) the legacy behavior for text filters is to return any argument
that is not a string as-is (thus, returning the same type as the
original value).

Add some additional test cases to make sure we don't re-introduce the
same problem in the future (assuming the base class function signature
doesn't change, of course).

[Fixes: #45]