Make moodle-plugin-ci validate pass again.
It insists on finding class filter_multilang2 in file filter.php.

The class aliasing construct is not recognized by the PHP parser used
by moodle-plugin-ci. But we don't need class aliasing at all, because
sub-classing works perfectly well in this case.

While at it, add a comment telling prospective code spelunkers where
to find the Real Thing (and why it's in there).