Fix parent - child language files generation.
While the standard multi-language plugin only creates the language
file for the child, we need to create both the child and the parent
files. The difference is due to the functions that both plugin use to
get the parent languages.

This was the reason why the unit test that tried to use the parent -
child relationship data failed unless we also specified the child
language in the block lang list: we simply didn't get the parent
language back when asking the string manager.

Now that we have fixed the language files generation, the unit test
works as it should.