Add support for configurable "parent language behaviour"
Now the plugin offers a setting to configure how the filter will
behave, with respect to parent languages, when processing a language
block.

This should help address issues #25 #42, but should be tested by the
affected parties before we can release a new version of the
plugin (hence the -alpha-1 release version).