Fix example for 'Always use parent languages, including en' option
Because we want to show the effect of including 'en' as a parent
language, we need to specify a language that is a "descendant"
language of 'en', not 'en' itself. In this case, given that we want to
show that both 'en_us' and 'en' would make the filter display the
language block, we need to use a language that is a "descendant" of
'en_us' too. Hence 'en_us_k12'.