Tell moodle-plugin-ci to ignore phpcs copyright boilerplate check
That phpcs codesniff looks for the standard Moodle copyritht
boilerplate. Given that we don't use the exact standard copyrith
boilerplate (see commit f3e2975929ce for the rationale), it fails for
no real reason.

Until a codesniff that is suitable for third party plugins is
available in moodle-plugin-ci, we are going to disable this one.